refactor: split nelson-data.py and tests into focused modules
nelson-data.py (2557 lines) and test_nelson_data.py (1759 lines)
exceeded the 800-line guideline. Split into 5 source modules and
3 test files while preserving backward compatibility:

- nelson-data.py: thin CLI entry point (287 lines)
- nelson_data_utils.py: shared I/O, validation, constants (261 lines)
- nelson_data_memory.py: cross-mission memory store (491 lines)
- nelson_data_lifecycle.py: mission lifecycle commands (825 lines)
- nelson_data_fleet.py: fleet intelligence + analytics (806 lines)
- conftest.py: shared test helpers (171 lines)
- test_nelson_data.py: lifecycle tests (620 lines)
- test_nelson_data_fleet.py: fleet/analytics/brief tests (719 lines)
- test_nelson_data_memory.py: memory store + I/O tests (303 lines)

CI updated to discover all test files via pytest directory discovery.
